Add Mathematical kind
