+++ fixtures/mathematical/sample.sxm
